Create SECURITY.md
Very. 5.1.x is currently supported with the newest security updates. 
Very. 5.0.x is now obsolete and not to be considered secure.
Ver 4.0.x is currently supported with the newest security updates
Ver 4.0 is now obsolete and not to be considered secure. 
You can report a bug or issue to esobwan@gmail.com